Add validation function to SubmitCustomizable components (#1788)
## Summary

This PR introduces a validation function to each customizable submit
button component, as part of the ongoing submit button customization
project. The purpose of this function is to enable merchants to check
the validation status of a component before initiating a payment. The
function returns a Boolean indicating the component's validation status
and also triggers the corresponding validation UI.

## Motivation

To provide merchants with the ability to verify the validation status of
a component prior to starting the payment process.

## Changelog

<new>

- [Added] Implemented `func validate() -> Bool` to retrieve the
validation status of the component. Impacted components:
  - `AbstractPersonalInformationComponent`
  - `GiftCardComponent`
  - `CardComponent`
  - `ACHDirectDebitComponent`
  - `OnlineBankingComponent`
  - `UPIComponent`
  - `BLIKComponent`
  - `SEPADirectDebitComponent`
  - `CashAppPayComponent`

</new>