Fix crash: duplicate items when rendering payment methods list (#1840)
# Summary
iOS18 release introduced new crash which didn't happen before.

When rendering payments list, in rare occasions we could have multiple
payment methods having same title and description.
Providing ListItem that is used to render the list uses title as part of
it's `Hashable` and `Equatable` conformance, resulting in having two
identical entries.
This list typically has few sections: stored payment methods and generic
payment methods. While any stored payment method always has an unique
identifier coming from the backend, the rest of the sections do not.

That's why this fix does not use any backend-provided identifiers and
relies purely on `UUID` that is created upon `ListItem` instantiation
and doesn't get exposed anywhere else.
It can be improved by providing a parameter in the init so it could be
injected from the outside.

## Note:
`ListItem` has `identifier` property that can be mistreated easily and
abused. We use it so set an identifier that is later used as
`accessibilityIdentifier` for UI elements. This is misleading as for
such a case when we have two identical stored payment methods this could
be a potential accessibility problem that will make second duped element
on the list inaccessible.

## Screenshot 

![image](https://github.com/user-attachments/assets/15f62807-30e7-46a7-b786-fcb11e2f1a0b)

<youtrack>COIOS-797</youtrack>