Expose `AdyenCardScanner` for dependency managers (#2097)
# Summary

This is the part of exploration on how to make `canImport` compile time
check work in the app that uses Cocoapods (MyStore).
The only way I found to produce a separate framework file is to expose
dependency as a separate pod.
This change:
- adds podspec for AdyenCardScanner 
- makes its code fully compatible with iOS12, otherwise no merchant
could build an iOS12 project with this framework
- replacing compile time canImport checks with some availability checks
(typealias CardScannerController = DummyCardScannerController) as a
result of approach change
- exposing `AdyenCardScanner` as SPM library

# Ticket

<ticket>
COIOS-894
</ticket>