Update PULL_REQUEST_TEMPLATE.md (#2092)
# Summary

We've been using the PR template for a few months now. Our recent
finding with mixed up release notes triggered this change.
I tried to rethink it in general and add more structure to it, following
our typical scenarios, adding more html-style comments as a guidance
that is not visible in the final markdown render.

This is an open discussion, comments are welcomed and appreciated.