v6: super component callbacks (#2119)
# Summary [Required]
The exposed new way of callbacks structure is through closures such as
onSubmit(), onAdditionalDetails(). Instead of changing this on component
level, we came up with the idea to have the new AdyenCheckoutComponent
be the bridge between the old apporach and the new by acting as the
delegate of components while using the closures for outside access.

## Ticket [Optional]
<ticket>
COSDK-467
</ticket>