CI - introduce SDK size checks (#2139)
# Summary
<!-- Provide a concise description (2-4 sentences) of what changes this
PR implements and why -->
The goal of this change is to increase SDK size visibility.
This is the first version that enumerates through all frameworks built
in release mode and outputs their sizes as a PR comment.

Changes: 
1. Added a new shell script to build the project and measure framework
sizes (Scripts/sdk_size.sh)
2. Added Github Action that posts framework sizes as PR comment if run
as a part of PR checks or calculates and stores target branch SDK sizes
if run manually

## Todo
<!-- Mark completed items with an [x] -->
- [x] Support caching so we can compare framework sizes
- [x] Post/update sizes as a PR comment
- [ ] Check comparison between base and target branches after this PR is
merged into develop (so we can run this manually on develop to capture
SDK size data)