COSDK-464: Introduce MVVM architecture in DropInComponent (#2143)
# Summary

This change introduces MVVM architecture to the `DropInComponent`. Now,
each one of the screens in the DropInComponent is its own MVVM module.
The idea is to move away from the `Component` pattern that we currently
have, and keep this one only for strict payment components.

The DropInComponent is full screen now, there's no usage of
`WrapperViewController` or `ModalViewController` to achieve partial
screens. Furthermore, the custom presentation is gone too.

# Demo



https://github.com/user-attachments/assets/ca0d2591-427a-4c4b-93db-2dec6842d806


## Ticket
<ticket>
COSDK-464
</ticket>

## Checklist [Required]
<!-- Mark completed items with an [x] -->
- [ ] Tested changes locally
- [ ] Added/updated unit tests
- [ ] Verified against acceptance criteria

---------

Co-authored-by: Andrii Mamchenko <atmamont@gmail.com>