blik component v6 example on advanced flow (#2202)
# Summary [Required]
AdyenCheckout and CheckoutConfiguration should be ready to be
initialized.
So it's a good time to start the demo sample code for Blik component and
see if there is anything missing to prevent us from a deliverable BLIK
component.

- [x] Create BLIKComponentExample file on the session or advanced flow
(both need to be done eventually)
   - [ ] Session flow
   - [x] Advanced flow
- [x] Instantiate CheckoutConfig with BLIK configuration
- [x] Supply onSubmit and onAdditionalDetails if advanced flow (can also
be used on session flow if they are passed)
- [x] Instantiate AdyenCheckout and pass the configuration
- [x] Present the component
- [ ] Error handling

## Ticket [Optional]
<ticket>
COSDK-576
</ticket>

## Checklist [Required]
<!-- Mark completed items with an [x] -->
- [x] Tested changes locally
- [ ] Added/updated unit tests
- [x] Verified against acceptance criteria