Improve unit/integration tests for `AdyenTheme` (#2318)
# Summary [Required]
<!-- Provide a concise description (2-4 sentences) of what changes this
PR implements and why -->

Moving fast means either no tests or right tests. So far unit tests
added to cover the `AdyenTheme` functionality were focusing primarily on
it's public API and builder methods testing.
Since we decided to shrink the `AdyenTheme` API quite a bit and expose
colors only, some changes were made.

A lot of unit tests were broken, and it takes time to update them
accordingly every time. It doesn't feel like the right thing to spend a
lot of time on.

Instead, I decide to change a testing approach a bit - for fast-changing
pieces like `AdyenTheme` it makes less sense to capture its internal
structure such as `AdyenLabelStyle` and so on and focus more on the
outcomes.

This PR deleted all unit tests that capture public API and internal
implementation details, and keeps the focus on integration tests, that
follow the structure:

**"Change the public API input such as color palette, and then verify if
the UI layer responds accordingly"**

This change should be merged before this
https://github.com/Adyen/adyen-ios/pull/2317