`AdyenTheme`: collect default values in one place (#2317)
# Summary [Required]
<!-- Provide a concise description (2-4 sentences) of what changes this
PR implements and why -->

## Part 1
After we made a decision to not expose internal UI elements styling
options, we do not need to provide a flexible public API with lots of
builder methods and default fallbacks anymore.

One of the downsides of the prior approach was styling values spreaded
across multiple structs and their inits.
It is much easier to track theme changes and compare them between
platforms when they are applied in one single place, this change tries
to achieve this level of maintainability.

## Part 2
I noticed we do not pass the `AdyenTheme` to the `CardViewController`,
which is required for testing upcoming changes. This change adds proper
`theme` propagation.

Tests are updated in the next PR to make review easier.