Migrate `FormVerticalStackItemView` to `AdyenTheme` (#2327)
# Summary [Required]
<!-- Provide a concise description (2-4 sentences) of what changes this
PR implements and why -->

To re-enable tests that were previously commented out, I want to focus
on UI elements migration that could unlock this.

This PR migrates `FormVerticalStackItemView` and adds theme builder
methods to some views so tests could be re-enabled.

This PR re-enables the `CardComponentTests.testUIConfiguration` test,
which had to be rewritten to respect new `AdyenTheme` instead of the old
"Style" approach.
It is worth mentioning not everything could be tested after we shrinked
the public API of `AdyenTheme`. We cannon now, for instance, test:
- font and color change of each label
- individual text input background color
- per-field font size
- individual alignments
- individual toggle color

## Ticket [Optional]
<ticket>
COSDK-775
</ticket>

## Checklist [Required]
<!-- Mark completed items with an [x] -->
- [x] Tested changes locally
- [x] Added/updated unit tests
- [x] Verified against acceptance criteria